React UI: Graph refactoring (#6382)
* move graph related files into own folder


* move graph helper functions into own file


* fix typo in file name


* fix typo in file name and lint fixes
